requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
decorator==4.0.11
idna==3.3
praw==7.6.0
prawcore==2.3.0
requests==2.28.1
six==1.10.0
update-checker==0.18.0
urllib3==1.26.11
websocket-client==1.3.3
lxml=4.9.1
PyYAML=6.0
